Refaktor: Renamet amplitude.ts til events.ts, og flyttet funksjonalitet som ikke har med events å gjøre.
Når vi kan skrive "import <funksjon> fra amplitude/events", så syntes jeg det blir tydeligere hva <funksjon> gjør, enn når vi skriver "import <funksjon> fra amplitude/amplitude"
