+++ src/Forside/Sammenligningspanel/Sammenligningspanel.tsx
@@ -1,14 +1,13 @@
 import React, { FunctionComponent, useRef } from 'react';
 import './Sammenligningspanel.less';
 import ReactToPrint from 'react-to-print';
-import {AlertStripeFeil, AlertStripeInfo} from 'nav-frontend-alertstriper';
+import { AlertStripeFeil } from 'nav-frontend-alertstriper';
 import { RestSummertSykefraværshistorikk } from '../../api/summertSykefraværshistorikk';
 import { RestStatus } from '../../api/api-utils';
-import { useSendEvent } from '../../amplitude/amplitude';
+import { useSendEvent } from '../../amplitude/events';
 import { Normaltekst, Systemtittel } from 'nav-frontend-typografi';
 import { RestAltinnOrganisasjoner } from '../../api/altinnorganisasjon-api';
 import { useOrgnr } from '../../utils/orgnr-hook';
-import Lenke from "nav-frontend-lenker";
 
 export const Sammenligningspanel: FunctionComponent<{
     restSummertSykefraværshistorikk: RestSummertSykefraværshistorikk;
