+++ src/amplitude/ekstradata.ts
@@ -17,6 +17,11 @@ import { mapTilPrivatElleOffentligSektor, Sektor } from '../utils/sektorUtils';
 import { RestSummertSykefraværshistorikk } from '../api/summertSykefraværshistorikk';
 import { getSammenligningResultatMedProsent } from '../Forside/vurdering-utils';
 import { SammenligningsType } from '../Forside/vurderingstekster';
+import { MutableRefObject, useContext, useEffect, useRef } from 'react';
+import { virksomhetsdataContext } from '../utils/virksomhetsdataContext';
+import { sykefraværshistorikkContext } from '../utils/sykefraværshistorikkContext';
+import { summertSykefraværshistorikkContext } from '../utils/summertSykefraværshistorikkContext';
+import { enhetsregisteretContext, EnhetsregisteretState } from '../utils/enhetsregisteretContext';
 
 export interface Ekstradata {
     næring2siffer: string;
@@ -34,7 +39,41 @@ export interface Ekstradata {
     sektor: Sektor;
 }
 
-export const getEkstraDataFraVirksomhetsdata = (
+export const useEkstraDataRef = (): MutableRefObject<Partial<Ekstradata>> => {
+    const restVirksomhetsdata = useContext<RestVirksomhetsdata>(virksomhetsdataContext);
+    const restSykefraværshistorikk = useContext<RestSykefraværshistorikk>(
+        sykefraværshistorikkContext,
+    );
+    const restSummertSykefraværshistorikk = useContext<RestSummertSykefraværshistorikk>(
+        summertSykefraværshistorikkContext,
+    );
+    const dataFraEnhetsregisteret = useContext<EnhetsregisteretState>(enhetsregisteretContext);
+
+    const ekstradata = useRef<Partial<Ekstradata>>({});
+
+    useEffect(() => {
+        ekstradata.current = {
+            ...getEkstraDataFraVirksomhetsdata(restVirksomhetsdata),
+            ...getEkstraDataFraSykefraværshistorikk(restSykefraværshistorikk),
+            ...getEkstraDataFraSummertSykefraværshistorikk(
+                restSummertSykefraværshistorikk,
+                restVirksomhetsdata,
+            ),
+            ...getEkstraDataFraEnhetsregisteret(
+                dataFraEnhetsregisteret.restOverordnetEnhet,
+                restVirksomhetsdata,
+            ),
+        };
+    }, [
+        restVirksomhetsdata,
+        dataFraEnhetsregisteret,
+        restSykefraværshistorikk,
+        restSummertSykefraværshistorikk,
+    ]);
+    return ekstradata;
+};
+
+const getEkstraDataFraVirksomhetsdata = (
     restVirksomhetsdata: RestVirksomhetsdata,
 ): Partial<Ekstradata> => {
     if (restVirksomhetsdata.status === RestStatus.Suksess) {
@@ -52,12 +91,12 @@ export const getEkstraDataFraVirksomhetsdata = (
     return {};
 };
 
-export const getEkstraDataFraSykefraværshistorikk = (
-    restSykefraværshistorikk: RestSykefraværshistorikk
+const getEkstraDataFraSykefraværshistorikk = (
+    restSykefraværshistorikk: RestSykefraværshistorikk,
 ): Partial<Ekstradata> => {
     if (restSykefraværshistorikk.status === RestStatus.Suksess) {
         const kvartalsvisSammenligning = konverterTilKvartalsvisSammenligning(
-            restSykefraværshistorikk.data
+            restSykefraværshistorikk.data,
         );
 
         const sammenligningForSisteKvartal = kvartalsvisSammenligning.pop();
@@ -76,7 +115,7 @@ export const getEkstraDataFraSykefraværshistorikk = (
     return {};
 };
 
-export const getEkstraDataFraEnhetsregisteret = (
+const getEkstraDataFraEnhetsregisteret = (
     restOverordnetEnhet: RestOverordnetEnhet,
     restvirksomhetsdata: RestVirksomhetsdata,
 ): Partial<Ekstradata> => {
@@ -106,7 +145,7 @@ const getAntallForskjelligeFarger = (...resultater: SykefraværVurdering[]): num
         : undefined;
 };
 
-export const getEkstraDataFraSummertSykefraværshistorikk = (
+const getEkstraDataFraSummertSykefraværshistorikk = (
     restSummertSykefraværshistorikk: RestSummertSykefraværshistorikk,
     restvirksomhetsdata: RestVirksomhetsdata,
 ): Partial<Ekstradata> => {
